Search time filter (#193)
