+++ requirements.txt
@@ -16,11 +16,10 @@ colorcet==2.0.2           # via hvplot
 coverage==5.0.3           # via panoptes-utils (setup.py), pytest-cov
 cycler==0.10.0            # via matplotlib
 decorator==4.4.1          # via mocket
-fastparquet==0.3.3        # via panoptes-utils (setup.py)
 flask==1.1.1              # via panoptes-utils (setup.py)
 google-api-core==1.16.0   # via google-cloud-bigquery, google-cloud-core
 google-auth==1.13.1       # via google-api-core, google-cloud-bigquery, google-cloud-storage
-google-cloud-bigquery[pandas]==1.24.0  # via panoptes-utils (setup.py)
+google-cloud-bigquery[pandas,pyarrow]==1.24.0  # via panoptes-utils (setup.py)
 google-cloud-core==1.3.0  # via google-cloud-bigquery, google-cloud-storage
 google-cloud-storage==1.28.0  # via panoptes-utils (setup.py)
 google-resumable-media==0.5.0  # via google-cloud-bigquery, google-cloud-storage
@@ -31,18 +30,16 @@ idna==2.9                 # via requests
 itsdangerous==1.1.0       # via flask
 jinja2==2.11.1            # via bokeh, flask
 kiwisolver==1.1.0         # via matplotlib
-llvmlite==0.31.0          # via numba
 loguru==0.4.1             # via panoptes-utils (setup.py)
 markdown==3.2.1           # via panel
 markupsafe==1.1.1         # via jinja2
 matplotlib==3.1.3         # via panoptes-utils (setup.py)
 mocket==3.8.4             # via panoptes-utils (setup.py)
 more-itertools==8.2.0     # via pytest
-numba==0.48.0             # via fastparquet
-numpy==1.18.1             # via astroplan, astropy, bokeh, fastparquet, holoviews, matplotlib, numba, pandas, panoptes-utils (setup.py), photutils, scipy
+numpy==1.18.1             # via astroplan, astropy, bokeh, holoviews, matplotlib, pandas, panoptes-utils (setup.py), photutils, pyarrow, scipy
 oauthlib==3.1.0           # via requests-oauthlib
 packaging==20.3           # via bokeh, pytest
-pandas==1.0.3             # via fastparquet, google-cloud-bigquery, holoviews, hvplot, panoptes-utils (setup.py)
+pandas==1.0.3             # via google-cloud-bigquery, holoviews, hvplot, panoptes-utils (setup.py)
 panel==0.9.5              # via holoviews
 param==1.9.3              # via colorcet, holoviews, panel, pyct, pyviz-comms
 pendulum==2.1.0           # via panoptes-utils (setup.py)
@@ -51,6 +48,7 @@ pillow==7.0.0             # via bokeh, panoptes-utils (setup.py)
 pluggy==0.13.1            # via pytest
 protobuf==3.11.3          # via google-api-core, google-cloud-bigquery, googleapis-common-protos
 py==1.8.1                 # via pytest
+pyarrow==0.17.0           # via google-cloud-bigquery
 pyasn1-modules==0.2.8     # via google-auth
 pyasn1==0.4.8             # via pyasn1-modules, rsa
 pycodestyle==2.6.0        # via panoptes-utils (setup.py)
@@ -75,8 +73,7 @@ ruamel.yaml.clib==0.2.0   # via ruamel.yaml
 ruamel.yaml==0.16.10      # via panoptes-utils (setup.py)
 scalpl==0.3.0             # via panoptes-utils (setup.py)
 scipy==1.4.1              # via panoptes-utils (setup.py)
-six==1.14.0               # via astroplan, cycler, fastparquet, google-api-core, google-auth, google-cloud-bigquery, google-resumable-media, mocket, packaging, protobuf, pytest-remotedata, python-dateutil, thrift, tweepy
-thrift==0.13.0            # via fastparquet
+six==1.14.0               # via astroplan, cycler, google-api-core, google-auth, google-cloud-bigquery, google-resumable-media, mocket, packaging, protobuf, pytest-remotedata, python-dateutil, tweepy
 tornado==6.0.4            # via bokeh
 tqdm==4.45.0              # via panel
 tweepy==3.8.0             # via panoptes-utils (setup.py)
